+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* [App] Moving product switcher to right of map as a panel, and improving per-product controls layout
+  [#99](https://gitlab.data.bas.ac.uk/MAGIC/SIIS/-/issues/99)
 * [App] Ability to toggle debug output
   [#98](https://gitlab.data.bas.ac.uk/MAGIC/SIIS/-/issues/98)